fix chart not showing issue
There was an issue in this branch where the SDWA Violations charts wouldn't show after drawing a box.

I think this was because of an unraised error in replacing the Y/N strings. When grouping the violation_type object, HEALTH_BASED becomes the index and should be referred to as such - it's the index whose strings we want to replace.